// src/app/interfaces/register-form.interface.ts
export interface RegisterForm {
    rol: string;
    email: string;
    password: string;
    password_2: string;
    nombre: string;
    apellidos: string;
    universidad: string;
    facultad: string;
    areaConocimiento: String;
    titulacion: string;
    sector: string;
    nombreSocioComunitario: String;
    terminos_aceptados: boolean;
    telefono: Number;
    url: string;
    mision: string;
}